requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
bleach==3.1.0
Django==3.0
django-autoslug==1.9.6
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-tempus-dominus==5.1.2.9
Pillow==6.2.1
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
Unidecode==1.1.1
webencodings==0.5.1
